// the Coterie of Clouds
## Checklist

The list below serves as a reminder to make sure, all needed changes were added to the repository. You can add a `x` between the square brackets to tick the items when in edit mode or just tick the checkboxes when not.

If you forgot something, don't worry. Just do the changes on the same branches and they are added automatically to the pull request.

- [x] Added changed map file
- [x] Used a descriptive pull request title
- [x] Wrote a short description of the change in the pull request description (see below)
- [x] Added a short sentence summarizing the changes to the `changelog.txt` file
- [X] Increased the number in `version.txt` by 1

## Included changes: Added the Coterie of Clouds on Yggdrasil, as well as the rebuilt Embassy.