Show map diffs in PRs (#85)
This adds https://github.com/Delwing/mudlet-map-diff-action which'll show a diff of the map on pull requests.

Currently uses my personal cloudinary account in lieu of anything better, happy to change that. 

Once we merge this and confirm it working in the Achaea repo, I'll replicate this workflow in other repositories.